+++ editor_patch/vtypes.h
@@ -6,5 +6,10 @@ namespace rf
     {
         int max_len;
         char* buf;
+
+        operator const char*() const
+        {
+            return buf;
+        }
     };
 }